COLORADO-2535 modifyDatastreamByReference Error
We were calling modifyDatastreamByReference and sending it a URL
that didn't necessarily exist depending on your drupal configuration.
I modified it so that it creates a file in a publically accessable
location. That way we can ingest it. Also added a line to delete
the file when we are done with it.